requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
pymongo
pandas
langchain
faiss-cpu
tiktoken
pytest
